+++ lib/terraspace/compiler/dsl/syntax/mod.rb
@@ -3,5 +3,7 @@ module Terraspace::Compiler::Dsl::Syntax
     include Terraspace::Util::Logging
     include_dir("mod")
     include_dir("helpers")
+    include_plugin_helpers
+    include_project_level_helpers
   end
 end